# Reuse slice for the range vector allocations.