Signed-off-by: Cyril Tovena <cyril.tovena@gmail.com>
```
❯ benchcmp old.txt new.txt
benchmark                         old ns/op      new ns/op      delta
BenchmarkRangeQuery100000-16      3356959        3097925        -7.72%
BenchmarkRangeQuery200000-16      15821777       14543104       -8.08%
BenchmarkRangeQuery500000-16      2107956639     1992137391     -5.49%
BenchmarkRangeQuery1000000-16     4051613799     4041557363     -0.25%

benchmark                         old allocs     new allocs     delta
BenchmarkRangeQuery100000-16      5241           5235           -0.11%
BenchmarkRangeQuery200000-16      5993           5835           -2.64%
BenchmarkRangeQuery500000-16      138631         130304         -6.01%
BenchmarkRangeQuery1000000-16     272021         255363         -6.12%

benchmark                         old bytes     new bytes     delta
BenchmarkRangeQuery100000-16      716984        714652        -0.33%
BenchmarkRangeQuery200000-16      1464403       1338725       -8.58%
BenchmarkRangeQuery500000-16      135321968     132657776     -1.97%
BenchmarkRangeQuery1000000-16     270999096     265668520     -1.97%
```